refactor(cli): move chart archive extraction into pkg/chart
The zip-slip guard and chart-root lookup lived in the lint command, where
nothing else could reach them. The preinstall image check needs the same
unpacking, and a second copy of a security-sensitive path is how the two
drift apart.

Also cap total expanded size and entry count, so a hostile .tgz cannot fill
the disk of whatever unpacks it.
